+++ test/model.populate.test.js
@@ -3251,6 +3251,42 @@ describe('model: populate:', function() {
       db.close(done);
     });
 
+    it('populating an array of refs, slicing, and fetching many (gh-5737)', function(done) {
+      var BlogPost = db.model('gh5737_0', new Schema({
+        title: String,
+        fans: [{ type: ObjectId, ref: 'gh5737' }]
+      }));
+      var User = db.model('gh5737', new Schema({ name: String }));
+
+      User.create([{ name: 'Fan 1' }, { name: 'Fan 2' }], function(error, fans) {
+        assert.ifError(error);
+        var posts = [
+          { title: 'Test 1', fans: [fans[0]._id, fans[1]._id] },
+          { title: 'Test 2', fans: [fans[1]._id, fans[0]._id] }
+        ];
+        BlogPost.create(posts, function(error) {
+          assert.ifError(error);
+          BlogPost.
+            find({}).
+            slice('fans', [0, 5]).
+            populate('fans').
+            exec(function(err, blogposts) {
+              assert.ifError(error);
+
+              assert.equal(blogposts[0].title, 'Test 1');
+              assert.equal(blogposts[1].title, 'Test 2');
+
+              assert.equal(blogposts[0].fans[0].name, 'Fan 1');
+              assert.equal(blogposts[0].fans[1].name, 'Fan 2');
+
+              assert.equal(blogposts[1].fans[0].name, 'Fan 2');
+              assert.equal(blogposts[1].fans[1].name, 'Fan 1');
+              done();
+            });
+        });
+      });
+    });
+
     it('maps results back to correct document (gh-1444)', function(done) {
       var articleSchema = new Schema({
         body: String,
@@ -4443,6 +4479,104 @@ describe('model: populate:', function() {
           then(function(b) { return B.findById(b).populate('a').exec(); }).
           then(function(b) {
             assert.equal(b.toObject().a[0].name, 'test');
+            done();
+          }).
+          catch(done);
+      });
+
+      it('with functions for localField and foreignField (gh-5704)', function(done) {
+        var ASchema = new Schema({
+          name: String
+        });
+
+        var BSchema = new Schema({
+          name: String,
+          localField: String,
+          firstId: ObjectId,
+          secondId: ObjectId
+        }, {
+          toObject: { virtuals: true },
+          toJSON:   { virtuals: true }
+        });
+
+        BSchema.virtual('a', {
+          ref: 'gh5704',
+          localField: function() { return this.localField; },
+          foreignField: function() { return '_id'; },
+          justOne: true
+        });
+
+        var A = db.model('gh5704', ASchema);
+        var B = db.model('gh5704_0', BSchema);
+
+        A.create([{ name: 'test1' }, { name: 'test2' }]).
+          then(function(arr) {
+            return B.create([
+              {
+                name: 'b1',
+                localField: 'firstId',
+                firstId: arr[0]._id,
+                secondId: arr[1]._id
+              },
+              {
+                name: 'b2',
+                localField: 'secondId',
+                firstId: arr[0]._id,
+                secondId: arr[1]._id
+              }
+            ]);
+          }).
+          then(function() {
+            return B.find().populate('a').sort([['name', 1]]).exec();
+          }).
+          then(function(bs) {
+            assert.equal(bs[0].a.name, 'test1');
+            assert.equal(bs[1].a.name, 'test2');
+            done();
+          }).
+          catch(done);
+      });
+
+      it('with functions for ref (gh-5602)', function(done) {
+        var ASchema = new Schema({
+          name: String
+        });
+
+        var BSchema = new Schema({
+          referencedModel: String,
+          aId: ObjectId
+        });
+
+        BSchema.virtual('a', {
+          ref: function() { return this.referencedModel; },
+          localField: 'aId',
+          foreignField: '_id',
+          justOne: true
+        });
+
+        var A1 = db.model('gh5602_1', ASchema);
+        var A2 = db.model('gh5602_2', ASchema);
+        var B = db.model('gh5602_0', BSchema);
+
+        A1.create({ name: 'a1' }).
+          then(function(a1) {
+            return A2.create({ name: 'a2' }).then(function(res) {
+              return [a1].concat(res);
+            });
+          }).
+          then(function(as) {
+            return B.create([
+              { name: 'test1', referencedModel: 'gh5602_1', aId: as[0]._id },
+              { name: 'test2', referencedModel: 'gh5602_2', aId: as[1]._id }
+            ]);
+          }).
+          then(function() {
+            return B.find().populate('a').sort([['name', 1]]);
+          }).
+          then(function(bs) {
+            assert.equal(bs.length, 2);
+            assert.equal(bs[0].a.name, 'a1');
+            assert.equal(bs[1].a.name, 'a2');
             done();
           }).
           catch(done);